+++ src/Filterable.ts
@@ -24,7 +24,7 @@ import {
   separate
 } from './Compactable'
 import { Either } from './Either'
-import { pipe, Refinement } from './function'
+import { pipe } from './function'
 import {
   Functor,
   Functor1,
@@ -45,6 +45,7 @@ import {
 import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 import { getLeft, getRight, Option } from './Option'
 import { Predicate } from './Predicate'
+import { Refinement } from './Refinement'
 import { separated, Separated } from './Separated'
 
 // -------------------------------------------------------------------------------------